Rename captkExampleAlgorithm to CaPTkExampleAlgorithm
