+++ Plugins/upenn.cbica.captk.example/src/internal/QmitkCaPTkExampleView.cpp
@@ -30,7 +30,7 @@
 #include <QMessageBox>
 
 // CaPTk
-#include "captkExampleAlgorithm.h"
+#include "CaPTkExampleAlgorithm.h"
 
 #include "tinyxml.h"
